Refaktorering: Legger kode knytta til et spesifikt API i tilhørende fil for det API-et.
Dette gir hakket mer konsistent og domenebasert kode.
