+++ src/Kalkulator/Kalkulator/KalkulatorMedDagsverk.tsx
@@ -3,12 +3,12 @@ import './Kalkulator.less';
 import Kostnad from './../Kostnad/Kostnad';
 import { RestStatus } from '../../api/api-utils';
 import EksternLenke from '../../felleskomponenter/EksternLenke/EksternLenke';
-import { RestSykefraværshistorikk } from '../../api/kvartalsvisSykefraværshistorikk';
+import { RestSykefraværshistorikk } from '../../api/kvartalsvis-sykefraværshistorikk-api';
 import {
     getAntallTapteDagsverkSiste4Kvartaler,
     getKostnadForAntallDagsverk,
     Kalkulatorvariant,
-} from './../kalkulator-utils';
+} from '../kalkulator-utils';
 import { useSendEvent } from '../../amplitude/events';
 import { Kalkulatorrad } from './Kalkulatorrad/Kalkulatorrad';
