+++ src/mocking/generering-av-historikk-mock.ts
@@ -1,5 +1,5 @@
 import { ÅrstallOgKvartal } from '../utils/sykefraværshistorikk-utils';
-import { KvartalsvisSykefraværsprosent } from '../api/kvartalsvisSykefraværshistorikk';
+import { KvartalsvisSykefraværsprosent } from '../api/kvartalsvis-sykefraværshistorikk-api';
 
 const neste = (årstallOgKvartal: ÅrstallOgKvartal): ÅrstallOgKvartal => {
     const { årstall, kvartal } = årstallOgKvartal;
