+++ src/mocking/sykefraværshistorikk-mock.ts
@@ -1,7 +1,7 @@
 import {
     KvartalsvisSykefraværshistorikk,
     SykefraværshistorikkType,
-} from '../api/kvartalsvisSykefraværshistorikk';
+} from '../api/kvartalsvis-sykefraværshistorikk-api';
 import { genererHistorikk, genererMaskertHistorikk } from './generering-av-historikk-mock';
 
 const lagHistorikkMedLandOgSektor = (): KvartalsvisSykefraværshistorikk[] => {
